pyproject.toml
[project]
name = "mnist-led"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
